requirements.txt
Jinja2>=2.7.3
Markdown>=2.4.1
MarkupSafe>=0.23
Pygments>=1.6
Unidecode>=0.04.16
beautifulsoup4>=4.3.2
blinker>=1.3
docutils>=0.11
feedgenerator>=1.7
feedparser>=5.1.3
lxml>=3.3.5
pelican>=3.3
pypandoc>=0.8.2
python-dateutil>=2.2
pytz>=2014.4
six>=1.7.2
smartypants>=1.8.3
typogrify>=2.0.4
